updated fan control with 8 steps starting from 20c
updated fan control with 8 steps starting from 20c, when using the MAX31855K and when it is not mounted close/on the main board, the mainboard gets quite warm because it is quite close to the insulation/bottom case. the temp of the cold junction of MAX31855K does not get hot enough to start the fan. this will make the fan start earlier at a lower rate. starts at roughly 12.5% from 20c and increases by roughly 12.5% at each step